Move a few things out of runtime-common that were only consumed by host
