--- packages/runtime-common/diff-entries.ts
@@ -1,31 +1,0 @@
-function changedEntry(
-  listings: { path: string; lastModified?: number }[],
-  entry: { path: string; lastModified?: number }
-) {
-  return listings.some(
-    (item) =>
-      item.path === entry.path && item.lastModified != entry.lastModified
-  );
-}
-
-function hasEntry(
-  listings: { path: string; lastModified?: number }[],
-  entry: { path: string; lastModified?: number }
-) {
-  return listings.some((item) => item.path === entry.path);
-}
-
-export function diff(
-  prevEntries: { path: string; lastModified?: number }[],
-  currEntries: { path: string; lastModified?: number }[]
-) {
-  let changed = prevEntries.filter((entry) => changedEntry(currEntries, entry));
-  let added = currEntries.filter((entry) => !hasEntry(prevEntries, entry));
-  let removed = prevEntries.filter((entry) => !hasEntry(currEntries, entry));
-
-  return {
-    added: added.map((e) => e.path),
-    removed: removed.map((e) => e.path),
-    changed: changed.map((e) => e.path),
-  };
-}
